+++ src/results/App.svelte
@@ -36,17 +36,15 @@
     Status,
     CONF_DIRECTORY,
   } from './types'
-  import {
-    TreeType,
-    CallTraceFunction,
-    EventTypesFromExtension,
-    Expandable,
-  } from './types'
+  import { TreeType, CallTraceFunction, EventTypesFromExtension } from './types'
   import NewRun from './components/NewRun.svelte'
 
   import { writable } from 'svelte/store'
-  import { expandables } from './store/store'
-  import { element } from 'svelte/internal'
+  import {
+    expandables,
+    expandCollapse,
+    verificationResults,
+  } from './store/store'
 
   export const hide = writable([])
   export const pos = writable({ x: 0, y: 0 })
@@ -55,7 +53,6 @@
   let outputRunName: string
   let selectedCalltraceFunction: CallTraceFunction
 
-  let verificationResults: Verification[] = []
   let runningScripts: { pid: number; confFile: string; uploaded: boolean }[] =
     []
   let runs: Run[] = []
@@ -64,6 +61,11 @@
   let namesMap: Map<string, string> = new Map()
   let runsCounter = 0
   let focusedRun: string = ''
+
+  // listen to the results array to see if there are results or not
+  $: $verificationResults.length > 0
+    ? ($expandCollapse.hasResults = true)
+    : ($expandCollapse.hasResults = false)
 
   function newFetchOutput(e: CustomEvent<Assert | Rule>, vr: Verification) {
     let clickedRuleOrAssert = e.detail
@@ -106,6 +108,32 @@
     if (selectedCalltraceFunction) selectedCalltraceFunction = undefined
   }
 
+  /**
+   * updated the results tree expand values according to results
+   */
+  function updateExpendablesFromResults() {
+    $verificationResults.forEach(vr => {
+      return vr.jobs.forEach(job => {
+        const temp = job.verificationProgress.rules.map(rule => {
+          return {
+            title: rule.name,
+            isExpanded: false,
+            tree: [],
+          }
+        })
+        $expandables = $expandables.map(element => {
+          if (
+            element.title === namesMap.get(job.runName) &&
+            element.tree.length === 0
+          ) {
+            element.tree = temp
+          }
+          return element
+        })
+      })
+    })
+  }
+
   const listener = (e: MessageEvent<EventsFromExtension>) => {
     switch (e.data.type) {
       case EventTypesFromExtension.ReceiveNewJobResult: {
@@ -113,7 +141,7 @@
           action: 'Smart merge current results with new result',
           source: Sources.ResultsWebview,
           info: {
-            currentVerificationResults: verificationResults,
+            currentverificationResults: $verificationResults,
             newResult: e.data.payload,
             name: e.data.payload.runName,
           },
@@ -123,47 +151,13 @@
           e.data.payload.verificationReportLink,
         )
         smartMergeVerificationResult(
-          verificationResults,
+          $verificationResults,
           e.data.payload,
           e.data.payload.runName,
         )
-        verificationResults = verificationResults
+        $verificationResults = $verificationResults
 
-        verificationResults.forEach(vr => {
-          return vr.jobs.forEach(job => {
-            const temp = job.verificationProgress.rules.map(rule => {
-              return {
-                title: rule.name,
-                isExpanded: false,
-                tree: [],
-              }
-            })
-            $expandables = $expandables.map(element => {
-              if (
-                element.title === namesMap.get(job.runName) &&
-                element.tree.length === 0
-              ) {
-                element.tree = temp
-              }
-              return element
-            })
-          })
-        })
-
-        // console.log('testttt:', ruleObjArr)
-
-        // $expandables = $expandables.map(function(part, index, arr) {
-        //   for (let i = 0; i < ruleObjArr.length; i++) {
-        //     for (let j = 0; j < ruleObjArr[i].length; j++){
-        //       if (ruleObjArr[i][j].inherit === part.title) {
-        //         arr[index].tree = ruleObjArr[i][j].treeArr
-        //       }
-        //     }
-        //   }
-        //   return arr[index]
-        // });
-
-        console.log('expandanles from verification update:', $expandables)
+        updateExpendablesFromResults()
 
         if (e.data.payload.jobStatus === 'SUCCEEDED') {
           if (e.data.payload.runName) {
@@ -175,7 +169,7 @@
           action: 'After Smart merge current results with new result',
           source: Sources.ResultsWebview,
           info: {
-            updatedVerificationResults: verificationResults,
+            updated$verificationResults: $verificationResults,
           },
         })
         break
@@ -235,7 +229,7 @@
         const curRun = runs.find(run => run.id === pid)
         if (curRun !== undefined) {
           if (
-            verificationResults.find(vr => vr.name === curRun.name) !==
+            $verificationResults.find(vr => vr.name === curRun.name) !==
             undefined
           ) {
             runs = setStatus(curRun.name, Status.unableToRun)
@@ -290,10 +284,10 @@
           action: 'Received "clear-all-jobs" command',
           source: Sources.ResultsWebview,
           info: {
-            currentVerificationResults: verificationResults,
+            current$verificationResults: $verificationResults,
           },
         })
-        if (verificationResults.length > 0) verificationResults = []
+        if ($verificationResults.length > 0) $verificationResults = []
         clearOutput()
         break
       }
@@ -473,8 +467,6 @@
         run.status = Status.missingSettings
       }
       runsCounter = runs.push(run)
-      console.log(run.name)
-      console.log(namesMap.get(run.name))
       addNewExpendable(namesMap.get(run.name))
     } else {
       // don't create more than one new run while in rename state
@@ -504,7 +496,7 @@
     const name = runToDelete.name
 
     //delete results
-    verificationResults = verificationResults.filter(vr => {
+    $verificationResults = $verificationResults.filter(vr => {
       return vr.name !== name
     })
 
@@ -565,7 +557,7 @@
     // rename existing run
     if (oldName !== '') {
       // the renamed run should have the same verification results, if they exist
-      let oldResult = verificationResults.find(vr => vr.name === oldName)
+      let oldResult = $verificationResults.find(vr => vr.name === oldName)
       if (oldResult !== undefined) {
         let newResult: Verification = {
           name: newName,
@@ -573,10 +565,10 @@
           spec: oldResult.spec,
           jobs: oldResult.jobs,
         }
-        verificationResults = verificationResults.filter(vr => {
+        $verificationResults = $verificationResults.filter(vr => {
           return vr.name !== oldName
         })
-        verificationResults.push(newResult)
+        $verificationResults.push(newResult)
       }
 
       const oldConfNameMap: JobNameMap = {
@@ -610,7 +602,7 @@
     if (pendingQueue.length > 0) {
       let curRun = pendingQueue.shift()
       pendingQueueCounter--
-      verificationResults = verificationResults.filter(vr => {
+      $verificationResults = $verificationResults.filter(vr => {
         return vr.name !== curRun.fileName
       })
       runScript(curRun)
@@ -658,7 +650,7 @@
     pendingQueue = pendingQueue.filter(rq => {
       return rq.fileName !== run.name
     })
-    verificationResults = verificationResults.filter(vr => {
+    $verificationResults = $verificationResults.filter(vr => {
       return vr.name !== run.name
     })
     pendingQueueCounter--
@@ -677,8 +669,36 @@
     askToDeleteJob(jobNameMap)
   }
 
+  /**
+   * operates the expand / collapse functionality
+   */
+  function expandCollapseAll() {
+    if ($expandCollapse.hasResults) {
+      $expandables = $expandables.map(element => {
+        if (element.title !== 'JOB LIST') {
+          element.isExpanded = $expandCollapse.var
+        }
+        element.tree = element.tree.map(treeItem => {
+          treeItem.isExpanded = $expandCollapse.var
+          return treeItem
+        })
+        return element
+      })
+      $expandCollapse.title = $expandCollapse.var
+        ? 'Collapse All'
+        : 'Expand All'
+      $expandCollapse.icon = $expandCollapse.var ? 'collapse-all' : 'expand-all'
+      $expandCollapse.var = !$expandCollapse.var
+    }
+  }
+
   onMount(() => {
     window.addEventListener('message', listener)
+    $expandables.push({
+      title: 'JOB LIST',
+      isExpanded: true,
+      tree: [],
+    })
     if (runs.length === 0) {
       initResults()
     }
@@ -729,7 +749,6 @@
   <div>
     <Pane
       title="JOB LIST"
-      initialExpandedState={true}
       fixedActions={[
         {
           title: 'Run All',
@@ -745,6 +764,12 @@
           title: 'Create New Job',
           icon: 'diff-added',
           onClick: createRun,
+        },
+        {
+          title: $expandCollapse.title,
+          icon: $expandCollapse.icon,
+          onClick: expandCollapseAll,
+          disabled: !$expandCollapse.hasResults,
         },
       ]}
     >
@@ -767,7 +792,6 @@
                 duplicateFunc={duplicateRun}
                 runFunc={() => run(runs[index])}
                 status={runs[index].status}
-                {verificationResults}
                 {newFetchOutput}
                 nowRunning={(runningScripts.find(
                   rs => getFilename(rs.confFile) === runs[index].name,
@@ -775,23 +799,20 @@
                   (pendingQueue.find(rs => rs.fileName === runs[index].name) !==
                     undefined &&
                     pendingQueueCounter > 0)) &&
-                  verificationResults.find(
+                  $verificationResults.find(
                     vr => runs[index].name === vr.name,
                   ) === undefined}
                 isPending={pendingQueue.find(
                   rs => rs.fileName === runs[index].name,
                 ) !== undefined && pendingQueueCounter > 0}
-                expandedState={verificationResults.find(
+                expandedState={$verificationResults.find(
                   vr => vr.name === runs[index].name,
                 ) !== undefined}
-                initialExpandedState={$expandables.find(element => {
-                  return element.title === namesMap.get(runs[index].name)
-                })?.isExpanded}
                 pendingStopFunc={() => {
                   pendingStopFunc(runs[index])
                 }}
                 runningStopFunc={() => {
-                  verificationResults = verificationResults.filter(vr => {
+                  $verificationResults = $verificationResults.filter(vr => {
                     return vr.name !== runs[index].name
                   })
                   runs = setStatus(runs[index].name, Status.ready)
@@ -816,7 +837,6 @@
   {#if output.variables && output.variables.length > 0}
     <Pane
       title={`${output.treeViewPath.ruleName} variables`}
-      initialExpandedState={true}
       actions={[
         {
           title: 'Close Output',
@@ -829,7 +849,7 @@
     </Pane>
   {/if}
   {#if output.callTrace && Object.keys(output.callTrace).length > 0}
-    <Pane title={`CALL TRACE`} initialExpandedState={true}>
+    <Pane title={`CALL TRACE`}>
       <Tree
         runDisplayName=""
         data={{
@@ -841,25 +861,19 @@
     </Pane>
   {/if}
   {#if selectedCalltraceFunction && selectedCalltraceFunction.variables && selectedCalltraceFunction.variables.length > 0}
-    <Pane
-      title={`${selectedCalltraceFunction.name} variables`}
-      initialExpandedState={true}
-    >
+    <Pane title={`${selectedCalltraceFunction.name} variables`}>
       <CodeItemList codeItems={selectedCalltraceFunction.variables} />
     </Pane>
   {/if}
   {#if output.callResolutionWarnings && output.callResolutionWarnings.length > 0}
-    <Pane
-      title={`Contract call resolution warnings`}
-      initialExpandedState={true}
-    >
+    <Pane title={`Contract call resolution warnings`}>
       {#each output.callResolutionWarnings as resolution}
         <ContractCallResolution contractCallResolution={resolution} />
       {/each}
     </Pane>
   {/if}
   {#if output.callResolution && output.callResolution.length > 0}
-    <Pane title={`Contract call resolution`} initialExpandedState={true}>
+    <Pane title={`Contract call resolution`}>
       {#each output.callResolution as resolution}
         <ContractCallResolution contractCallResolution={resolution} />
       {/each}