+++ src/results/types.ts
@@ -262,10 +262,3 @@ export type Run = {
   status: Status
   vrLink?: string
 }
-
-export type Expandable = {
-  title: string
-  isExpandable: boolean
-  tree: Expandable[]
-  inherit?: string
-}